+++ app/src/main/java/com/example/bcsd_weather/di/KoinModule.kt
@@ -4,17 +4,13 @@ package com.example.bcsd_weather.di
 import android.app.Application
 import androidx.room.Room
 import com.example.bcsd_weather.FutureWeatherViewModel
-import com.example.bcsd_weather.data.db.FutureWeatherDao
+import com.example.bcsd_weather.data.datasource.local.FutureWeatherDao
 import com.example.bcsd_weather.data.db.FutureWeatherDatabase
 import com.example.bcsd_weather.data.repository.FutureWeatherRepositoryImpl
-import com.example.bcsd_weather.db.future.converter.StItemTypeConverter
-import com.example.bcsd_weather.db.future.converter.shortTermForecastModelTypeConverter
 import com.example.bcsd_weather.domain.repository.FutureWeatherRepository
-import com.google.gson.Gson
 import org.koin.android.ext.koin.androidApplication
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
-import java.util.concurrent.Future
 
 
 val appModule = module {
