+++ data/src/main/java/com/example/bcsd_weather/data/db/FutureWeatherDatabase.kt
@@ -2,6 +2,7 @@ package com.example.bcsd_weather.data.db
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
+import com.example.bcsd_weather.data.datasource.local.FutureWeatherDao
 import com.example.bcsd_weather.data.model.FutureWeatherEntity
